Apply mypy-tests custom config to other mypy-based tests
