+++ tests/stubtest_third_party.py
@@ -16,6 +16,7 @@ from time import time
 from typing import NoReturn
 
 from ts_utils.metadata import NoSuchStubError, get_recursive_requirements, read_metadata
+from ts_utils.mypy import mypy_configuration_from_distribution, temporary_mypy_config_file
 from ts_utils.paths import STUBS_PATH, allowlists_path, tests_path
 from ts_utils.utils import (
     PYTHON_VERSION,
@@ -95,89 +96,95 @@ def run_stubtest(
             print_command_failure("Failed to install", e)
             return False
 
-        ignore_missing_stub = ["--ignore-missing-stub"] if stubtest_settings.ignore_missing_stub else []
-        packages_to_check = [d.name for d in dist.iterdir() if d.is_dir() and d.name.isidentifier()]
-        modules_to_check = [d.stem for d in dist.iterdir() if d.is_file() and d.suffix == ".pyi"]
-        stubtest_cmd = [
-            python_exe,
-            "-m",
-            "mypy.stubtest",
-            # Use --custom-typeshed-dir in case we make linked changes to stdlib or _typeshed
-            "--custom-typeshed-dir",
-            str(dist.parent.parent),
-            *ignore_missing_stub,
-            *packages_to_check,
-            *modules_to_check,
-            *allowlist_stubtest_arguments(dist_name),
-        ]
+        mypy_configuration = mypy_configuration_from_distribution(dist_name)
+        with temporary_mypy_config_file(mypy_configuration) as temp:
+            ignore_missing_stub = ["--ignore-missing-stub"] if stubtest_settings.ignore_missing_stub else []
+            packages_to_check = [d.name for d in dist.iterdir() if d.is_dir() and d.name.isidentifier()]
+            modules_to_check = [d.stem for d in dist.iterdir() if d.is_file() and d.suffix == ".pyi"]
+            stubtest_cmd = [
+                python_exe,
+                "-m",
+                "mypy.stubtest",
+                "--mypy-config-file",
+                temp.name,
+                # Use --custom-typeshed-dir in case we make linked changes to stdlib or _typeshed
+                "--custom-typeshed-dir",
+                str(dist.parent.parent),
+                *ignore_missing_stub,
+                *packages_to_check,
+                *modules_to_check,
+                *allowlist_stubtest_arguments(dist_name),
+            ]
 
-        stubs_dir = dist.parent
-        mypypath_items = [str(dist)] + [str(stubs_dir / pkg.name) for pkg in requirements.typeshed_pkgs]
-        mypypath = os.pathsep.join(mypypath_items)
-        # For packages that need a display, we need to pass at least $DISPLAY
-        # to stubtest. $DISPLAY is set by xvfb-run in CI.
-        #
-        # It seems that some other environment variables are needed too,
-        # because the CI fails if we pass only os.environ["DISPLAY"]. I didn't
-        # "bisect" to see which variables are actually needed.
-        stubtest_env = os.environ | {"MYPYPATH": mypypath, "MYPY_FORCE_COLOR": "1"}
+            stubs_dir = dist.parent
+            mypypath_items = [str(dist)] + [str(stubs_dir / pkg.name) for pkg in requirements.typeshed_pkgs]
+            mypypath = os.pathsep.join(mypypath_items)
+            # For packages that need a display, we need to pass at least $DISPLAY
+            # to stubtest. $DISPLAY is set by xvfb-run in CI.
+            #
+            # It seems that some other environment variables are needed too,
+            # because the CI fails if we pass only os.environ["DISPLAY"]. I didn't
+            # "bisect" to see which variables are actually needed.
+            stubtest_env = os.environ | {"MYPYPATH": mypypath, "MYPY_FORCE_COLOR": "1"}
 
-        # Perform some black magic in order to run stubtest inside uWSGI
-        if dist_name == "uWSGI":
-            if not setup_uwsgi_stubtest_command(dist, venv_dir, stubtest_cmd):
-                return False
+            # Perform some black magic in order to run stubtest inside uWSGI
+            if dist_name == "uWSGI":
+                if not setup_uwsgi_stubtest_command(dist, venv_dir, stubtest_cmd):
+                    return False
 
-        if dist_name == "gdb":
-            if not setup_gdb_stubtest_command(venv_dir, stubtest_cmd):
-                return False
+            if dist_name == "gdb":
+                if not setup_gdb_stubtest_command(venv_dir, stubtest_cmd):
+                    return False
 
-        try:
-            subprocess.run(stubtest_cmd, env=stubtest_env, check=True, capture_output=True)
-        except subprocess.CalledProcessError as e:
-            print_time(time() - t)
-            print_error("fail")
+            try:
+                subprocess.run(stubtest_cmd, env=stubtest_env, check=True, capture_output=True)
+            except subprocess.CalledProcessError as e:
+                print_time(time() - t)
+                print_error("fail")
 
-            print_divider()
-            print("Commands run:")
-            print_commands(pip_cmd, stubtest_cmd, mypypath)
+                print_divider()
+                print("Commands run:")
+                print_commands(pip_cmd, stubtest_cmd, mypypath)
 
-            print_divider()
-            print("Command output:\n")
-            print_command_output(e)
+                print_divider()
+                print("Command output:\n")
+                print_command_output(e)
 
-            print_divider()
-            print("Python version: ", end="", flush=True)
-            ret = subprocess.run([sys.executable, "-VV"], capture_output=True, check=False)
-            print_command_output(ret)
-
-            print("\nRan with the following environment:")
-            ret = subprocess.run([pip_exe, "freeze", "--all"], capture_output=True, check=False)
-            print_command_output(ret)
-            if keep_tmp_dir:
-                print("Path to virtual environment:", venv_dir, flush=True)
-
-            print_divider()
-            main_allowlist_path = allowlists_path(dist_name) / "stubtest_allowlist.txt"
-            if main_allowlist_path.exists():
-                print(f'To fix "unused allowlist" errors, remove the corresponding entries from {main_allowlist_path}')
-                print()
-            else:
-                print(f"Re-running stubtest with --generate-allowlist.\nAdd the following to {main_allowlist_path}:")
-                ret = subprocess.run([*stubtest_cmd, "--generate-allowlist"], env=stubtest_env, capture_output=True, check=False)
+                print_divider()
+                print("Python version: ", end="", flush=True)
+                ret = subprocess.run([sys.executable, "-VV"], capture_output=True, check=False)
                 print_command_output(ret)
 
-            print_divider()
-            print(f"Upstream repository: {metadata.upstream_repository}")
-            print(f"Typeshed source code: https://github.com/python/typeshed/tree/main/stubs/{dist.name}")
+                print("\nRan with the following environment:")
+                ret = subprocess.run([pip_exe, "freeze", "--all"], capture_output=True, check=False)
+                print_command_output(ret)
+                if keep_tmp_dir:
+                    print("Path to virtual environment:", venv_dir, flush=True)
 
-            print_divider()
+                print_divider()
+                main_allowlist_path = allowlists_path(dist_name) / "stubtest_allowlist.txt"
+                if main_allowlist_path.exists():
+                    print(f'To fix "unused allowlist" errors, remove the corresponding entries from {main_allowlist_path}')
+                    print()
+                else:
+                    print(f"Re-running stubtest with --generate-allowlist.\nAdd the following to {main_allowlist_path}:")
+                    ret = subprocess.run(
+                        [*stubtest_cmd, "--generate-allowlist"], env=stubtest_env, capture_output=True, check=False
+                    )
+                    print_command_output(ret)
 
-            return False
-        else:
-            print_time(time() - t)
-            print_success_msg()
-            if keep_tmp_dir:
-                print_info(f"Virtual environment kept at: {venv_dir}")
+                print_divider()
+                print(f"Upstream repository: {metadata.upstream_repository}")
+                print(f"Typeshed source code: https://github.com/python/typeshed/tree/main/stubs/{dist.name}")
+
+                print_divider()
+
+                return False
+            else:
+                print_time(time() - t)
+                print_success_msg()
+                if keep_tmp_dir:
+                    print_info(f"Virtual environment kept at: {venv_dir}")
     finally:
         if not keep_tmp_dir:
             rmtree(venv_dir)
